Added get_contact() function.
This function is called by the user when they leave the edit contact name box.  It retrieves the user by first and last name, and calls on the template manager function edit_contact_form_filler() to populate the form.  It supplies the retrieved data, and a list of the inputs to fill.