Update Contact functions.
Added the edit_contact to public API of contactService, and the update_contact function to the private method section.  Edit_contact retrieves all info from the fields, and calls on update_contact to peform th PUT request.  Update contact expects a callback function which is the contact_update_form_reset in ContactTemplate.js